18.2. Parallel AJAX with Promises
Promise.all: Parallel Promises
But look back at the watcher function. I bet a lot of you spotted something wrong here: it's inefficient! We're making the first AJAX call, waiting and then starting the second AJAX call. Sometimes you do need to wait for one AJAX call to finish before you start another one... because the second one depends on the first. But that is not the case here: we should be able to start both of these at the same time so they run in parallel.
We could accomplish that by refactoring these into separate methods... or with the .then() syntax: using the response argument to set the products data. We could do the same for colors.
But... I'm going to undo that. As a challenge, let's pretend that both of these AJAX calls need to finish before we can run either of these lines of code. Like, maybe because we need to combine the data from both endpoints in some way.
So the question is: how can we start two AJAX calls at the same time, but then wait until both of them finish? The answer is with a cool Promise.all() function.
Check it out: say const and then an array with productResponse and colorResponse inside. Set that equal to await Promise.all(). Pass this the two promises that we need to wait for: fetchProductsbyId() and... after some reorganizing... fetchColors().
I love this! Promise.all() takes an array of Promises and returns a Promise. That Promise resolves once all of the internal promises resolve. The final resolved data is an array... and so we use array destructuring to set the colorResponse and productResponse variables. That's some fancy JavaScript!
And when we try it... hey! That fanciness even works!