18.3. Parallel AJAX with Promises Loading Data Event Earlier But... as cool & hipster as this is, we don't really need it in this case. And we can make this even more performant. Think about it: because we're querying for all of the colors - and not just the colors for the items in the cart - we don't need to wait for the cart AJAX call to finish before fetching the colors. Nope, the color AJAX call can start immediately when the component is created. At the bottom, add an async created(). Inside, say this.colors = await, then move the fetchColors() here... but also with the .data['hydra:member'] on it. Oh, and make sure this is called created not create. Thanks to this, up here when the cart changes - so after the cart AJAX call finishes - we only need to make the one AJAX call for products. I'll simplify this again: const productResponse = await, then fetchProductsById(). Once that's finished, set the products data. Oh, and don't forget your equal sign! Let's make sure I didn't bust anything. Do a full refresh and... got it! Behind the scenes, the cart and colors AJAX calls both start immediately. Then, when the cart API call finishes, the products AJAX call starts. We can see this down in the browser network tools, filtered to the XHR calls. The cart call starts first, then colors - though, you can see on the waterfall on the right that they started at almost the same moment. Then, later, after the cart call finishes, here is the products AJAX call. Pretty cool.
Method Refactoring
Oh, but before we keep going, I want to make one tiny change. Add a methods key at the bottom of the component with one method inside: async loadProducts()
Move the three product AJAX lines from the cart watcher into this. Then, call it from the watcher: this.loadProducts().
There's no specific reason for this change. For me, it's more readable to give these three lines of code, a name, like loadProducts. It's now easier to understand that, when the cart changes, we load the products. I do this a lot in PHP by creating private methods.
Next: we have all the data we need! So let's make this page shine.