20.1. Inventing a Proper v-for :key
Our cart items are kind of weird... but it's not our fault! We fetch the cart via AJAX... but each item does not have a unique ID. Because of this, when we loop over the items in index.vue, we have to use the array index as the key.

Why do we care about :key Again?
In a more perfect, sunnier world with rainbows, the key would be something that's unique and will always be the same for each item. The array index is unique, but if we re-ordered the items inside the cart data, suddenly each would render with a different key than it had before.
What's the problem with that? Well... there may be no problem! But in certain situations, this could could cause some rows to re-render incorrectly. And... that's really the whole point of the key prop! To help Vue keep track of which component - or HTML element - is associated with each item in the array.

Creating a New, Unqiue :key
So let's improve this! How? By adding a unique key ourselves to each item! Really, if you combine an item's product and color, that forms a unique key. It is valid to have the same product in the cart twice, with different colors.
In shopping-cart.vue, find the completeCart computed property. We have a lot of power here to add anything we want to each item, like an id!
Change the function to use multiple lines instead of the super short format with the implied return statement: we need a bit more space to work. This is now equivalent to what we had before... except that ESLint is mad because it wants me to use the short format since we only have a return statement.
Add a new statement: const product = set to the products.find() line. Then... use that below: product: product. Or, even better, shorten it to product
Repeat this for color: const color =, copy the find() line, replace it with color, and paste that above
Cool! Except... for the duplicate product and color variables. In the callback, change the name to productItem... and also colorItem.
So far, this is the same... just written with extra lines. But now we can say: id:, use fancy "ticks", then ${cartItem.product}.
The cartItem variable comes from the original AJAX data, which means the product property is an IRI string. Add an underscore and do the same for color: ${} then if cartItem.color, print cartItem.color else the string none... or an empty string: whatever you want.
Nice! Let's go use this in index.vue: use the simpler syntax - item in items - then set :key to id.