20.2. Inventing a Proper v-for :key
Rendering More Cart Stuff!
Ok, enough of that! I want to fill in this page with more stuff! In index.vue, I'll paste in some HTML. This creates a, sort of, "table" structure for the items with quantity & price columns. Move the v-for into the middle of this. So, we have a row of headers, each cart-item is also a row, and then there are columns for each bit of data.