20.3. Inventing a Proper v-for :key
And... now the page is taking shape! For the price, we need to total up the price and quantity of each item. Hey! That's a textbook case for a computed prop!
Back in the component, add a computed key with one method called totalPrice. This is boring... "math" stuff, so I'll paste in the logic. At the bottom, return formatPrice(total).
Head to the top to import that function: import formatPrice from @/helpers/format-price.
Ok! Now that we have a totalPrice computed prop, we can use it in the template: {{ totalPrice }}.
Let's check it. Nice! We can't see the price of each item yet... or the quantity... but this total is probably correct.