21.1. Accidentally Modifying Prop Objects
Let's render the rest of the details for each row.

Rendering the Color Square
Inside cart-item, start by adding some some structure - <div class="col-2"> - with the name inside. Then add another div with col-1 to hold the product color, if there is one. Do that with a <span class="color-square" />
Head down to the bottom to style this: .color-square with some CSS to make this a 25px square with rounded borders. I'm also going to add a :global after .component. We talked about this in the first Vue tutorial: because we're nesting the .color-square inside of .component, it's safe to add :global and will avoid an extra, unnecessary module prefix before color-square in the final CSS.
All we need to do now is add a background-color set to the dynamic color. Add a style attribute - actually a :style attribute - set to an object with a backgroundColor key. Remember: instead of background-color, Vue wants you to use camel-case names: it handles converting things. Set this to a string using our magic "ticks": #, ${} then, to keep my template readable, use a new computed prop: hexColor.
Copy that, then head down to define it: computed, hex(), then return, if this.item.color - which will either be null or an object - return this.item.color.hexColor else fff, which will be an invisible white box. You could also use transparent. Or, if there is no color, you could avoid rendering any color box with v-if.
Anyways, let's see how it looks. Very nice!