21.2. Accidentally Modifying Prop Objects
The Quantity Input
Next up is quantity. I want this to render an input box so the user can easily change the quantity. Add a div with class="col-3" so that it lines up with the headers, which are each col-3. Then say <input.
Let's think: we want the value of this input to be equal to the item.quantity prop. And when we change the input, we want that to update the item.quantity value. Whelp, that is exactly the job of v-model. Add v-model - with the cool .number so that Vue converts the input into a number - equals item.quantity. Finish with class="form-control" and type="number".
Oh, and to make sure the input is small. At the bottom, add an input style with width: 60px.
Let's try it! Back at the browser, it instantly shows the correct numbers!

Whoops! We Modified a Prop!
But... there's a problem with this... a big problem. Do you remember how we're never supposed to modify a prop? Well... we just did. My bad!
The item is passed to us as a prop. Then, thanks to v-model, we're modifying the quantity key on that prop. Yikes!
In reality, this isn't quite as bad as changing, or replacing an entire prop. Think about it: on the top-level ShoppingCart component... this completeCart computed property is what is passed down to the other components. And each object inside of the items array will be the exact same object in memory as the item prop in the cart-item component. So if we change the item.quantity key in cart-item, that will also change the quantity property on that object in the top-level ShoppingCart component.... because... those are literally the same object!
We can see this! Change our floppy disk quantity to 12. In the Vue dev tools, find the first ShoppingCartItem component. Its item.quantity value is, of course, 12... which makes sense because we just set that value directly.
Now go look at the top-level ShoppingCart component, find the completeCart computed property and open the first item. Hey! It has quantity 12!
So... if a prop is an object... is modifying a key on that object really a problem? After all, everything seems to stay "in sync" !
In some cases... maybe not. But, you're playing with fire. In our situation, it causes things to... well... get weird.
Change the quantity back to 10. Look at the Vue dev tools! The quantity prop did not update! In reality, it did update... but you can't see it until you click off of the ShoppingCart component and re-open it. There is quantity 10.
The problem is that this just isn't how Vue is supposed to work... even if we're "kind of" getting away with it. And so, weird things like this start to happen.
And the Vue dev tool is not the only weird thing. Our situation is especially strange because, by accident, we're literally changing the quantity of a computed property!
That means that the we are not updating the quantity on the actual "source of truth", which is the cart data. Find the cart data and open the first item. Yep, it still has quantity 15: the original quantity. That is not changing. So if the shopping cart re-rendered right now, it would completely wipe out the updated value on the computed property and replace it with the one from data.
In other words... we accidentally modified a prop by modifying a key of an object! And while you might sometimes get away with this, you're asking for trouble.