21.3. Accidentally Modifying Prop Objects
Replacing v-model with value=
For us, it means that we cannot use v-model in this situation. But... that's ok! We already know the correct solution! Change this to :value="item.quantity" so that it renders the correct value. In a few minutes, we'll add some code so that when the input changes, we emit an event, which is the proper way to notify a parent component that something has changed.

Rendering the Price
Before we do that, there are two more quick things I want to render. Add another <div class="col-3">, a $ then print a totalPrice computed property.
Next, import the formatPrice helper that we've been using... then add a totalPrice function under computed. Return formatPrice() of this.item.product.price times this.item.quantity.
Finally, up in the template, add one more column with a button, some styling classes and the word Remove. We'll make this actually do something in a few minutes.
Moment of truth! Find your browser. Yes! That looks wonderful.
Next: when the user changes the quantity, we need to update the cart data on the ShoppingCart component and save the changes back to the server. Let's do that with - of course - events!