22.2. Emitting up the Component Tree
Converting the Quantity to a Number
This is exactly the reason why we sometimes use v-model.number: it grabs the value, converts it into a number and then updates the data. We're not using v-model here, but we can do that same conversion manually. Wrap the input value with parseFloat()
Now when we change the quantity... yes! We have a number!

Emitting the Event up the Component Chain
So: unfortunately, we can't just listen to the new updateQuantity event from shopping-cart... because shopping-cart doesn't render cart-item directly: it's rendered by the shopping cart list component.
But it's no problem! We just need to do one extra step: listen to the event in ShoppingCartList and re-emit that event so that we can then listen to it from the top-level ShoppingCart.
Open index.vue and find where we render shopping-cart-item: here it is. Add @updateQuantity to listen to the new event. Normally, we set this to a method name... but we can also just... write code here! Re-emit the event with ()... and I'll keep the same name - updateQuantity. For the data... just use the same event! We have access to it here as $event
Check this out in the Vue dev tools: now when we change the quantity... yes! There are two events, one emitted from each component.
Finally, we can listen to this event in ShoppingCart and use it to update the data and save the new quantity back to the server. Let's do that next!