23.1. Quantity: Updating Data & Saving to the Server
When we change the quantity input, the ShoppingCartList component is now emitting an updateQuantity event. In ShoppingCart, let's listen to that!
Up in the template, find where we render shopping-cart-list. Then add @updateQuantity= and make this execute a new method called, how about, updateQuantity.
Copy that, go down to the methods section... here it is... and add that new function.

Array Destructing the Event Args
We know that this will receive an event object... and because we are the ones emitting that event, we know the object will have productId, colorId and quantity keys.
In the new method, let's use array destructuring on the first argument to grab those values and set them onto variables. Do it with {}, productId, colorId and quantity
Let's console.log() all 3 values to make sure everything is hooked up
At the browser... change the quantity on the inflatable sofa. Yes! The product IRI, color IRI and quantity do log.