24.1. More Mixin Logic: Updating Quantity in Header
In cart-item.vue, when the quantity changes, we're emitting an updateQuantity event. But... it occurred to me that we're passing too much data in the event... or really, we're passing more data than we need to.

Passing Less Data to the Event
Think about it: both productId and colorId come from this.item... which is passed to us as a prop... which means that our parent component already knows the productId and colorId that this component instance is tied to. Passing these back to our parent in the event data is... redundant!
On a philosophical level, all we should need to pass to updateQuantity is the new quantity.
This is a tiny detail, but let's clean this up. Copy both productId and colorId and then delete them from the event data.
Now go to index.vue. When we emit the event here, we do need to include productId and colorId so that ShoppingCart knows which item's quantity is being updated. We can add that info right here: call () and this time pass an object. Paste those keys - productId and colorId - change this.item to item... and then make sure to continue including quantity set to .quantity.
That's it! It worked fine before... and it works fine now... but this feels better.

Fixing the Cart Header
The one thing that is not working when we change the quantity is the cart header: it's still not updating. But we do have logic to handle this. Open shopping-cart.vue and find the updateQuantity() method.
We know that this component uses a mixin: get-shopping-cart.js. And that holds the code to update the shopping cart count in the header. We run this after adding a new product to the cart.
Ok then: if we isolated this document
          .getElementById('js-shopping-cart-items')
          .innerHTML = getCartTotalItems(this.cart).toString() into its own method, we could call that from inside of shopping-cart.vue when the quantity is updated. Let's do it! But wait, let's go even further. We have this nice mixin... whose job is to hold cart data and methods related to that data. So it makes sense to also include a method here to handle everything related to updating an item's quantity.