24.2. More Mixin Logic: Updating Quantity in Header
Adding More Logic to the Mixin
Check it out: in the mixin, add a new method called updateProductQuantity() with 3 arguments for the 3 pieces of info this needs: productId, colorId and quantity.
Now go to shopping-cart, copy the updateCartItemQuantity() function that updates the quantity and makes the AJAX call, and paste it here. When I did that, PhpStorm automatically added the import on top for me... though I don't love that syntax.
Anyways, now that we have this method, we can call it from our shopping-cart component. Remove the console.log() and just say this.updateProductQuantity() passing the same arguments, except that we don't need to pass this.cart: the mixin already has that.
Sweet! Just a little bit of reorganization so that more cart-related logic lives in the mixin. And if we try it... yup! It's not broken. We're awesome!