24.3. More Mixin Logic: Updating Quantity in Header
Thanks to this, updating the cart header will be even easier. Start by copying the header logic and scrolling down so we can create a new method. Call it updateCartHeaderTotal() and paste!
Now, very simply, at the end of addProductToCart() call this.updateCartHeaderTotal() and... repeat that in updateProductQuantity: this.updateCartHeaderTotal().
Oh, and we don't need to do this, but I'm going to add an await and then make the method async. This will now wait for that quantity AJAX call to finish and then update the quantity in the header.
Let's try it! I'll do a full page refresh to be safe. The cart header says 13. Increase a quantity... boom! Everything updates - including the cart header.
Next: let's hook up the last part of the cart functionality: the remove from cart button.