26.1. "Add to Cart Controls" Component
The marketing department wants us to add a "featured product" to the cart sidebar with the ability to add that item to the cart right from this page.
Let's go look at the product show page for our trusty inflatable sofa. The most complex part of this page - by far - is the "add to cart controls": the quantity box, color selector, button and all the functionality related to those. I'd really like to reuse all of this on the featured product sidebar.
And so, this is a perfect opportunity to isolate this chunk of HTML and functionality into a beautiful re-usable component.

Refactoring to a product-show/ Directory
Start inside the assets/components/ directory: create a new folder called product-show/ for organization. Then move the product-show.vue component into that directory... but rename it to index.vue so that our existing import statements still work. But... this change does confuse Webpack: it doesn't realize that it should stop looking for product-show.vue and start looking for product-show/index.vue.

Ht Control+C on Encore to stop it, and then restart it:

  yarn dev-server
Perfect!