27.1. Finishing the Cart Controls Component
We've successfully extracted most of the markup and functionality for the "add to cart" controls into this new component. Our template still references a few undefined methods, but thanks to the props and data we added to the component, all the variables should now be defined. And that means, we should be able to get this to render.

Using the Component
Back in index.vue, let's import this puppy: import CartAddControls from and I'll be lazy and just use ./cart-add-controls because they live in the same directory. Add this to components...
Then head up to the template: <cart-add-controls. Pass this the props it needs: :product set to product and then allowAddToCart. This should be true if the cart is done loading.
As a reminder, this component uses the get-shopping-cart mixin, which means it has a data called cart. We can use that here: allowAddToCart if cart ! == null.
Next add :add-to-cart-loading set to the addToCartLoading data - which also comes from the mixin - and :add-to-cart-success set to addToCartSuccess.
Done! Let's try it out! Move over to the browser and... yes! It renders! Well, yes, it does render, but Vue is totally mad because our template references some undefined methods. Time to fix those!